refactor packagemanager into submodule
