+++ belay/packagemanager/__init__.py
@@ -1,0 +1,1 @@
+from .group import Group, GroupConfig
